backend_v3.0/requirements.txt
fastapi
uvicorn
python-dotenv
aiomysql
redis
pydantic
requests
